+++ src/main/java/com/profect/tickle/domain/event/service/application/CouponService.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.event.service.event;
+package com.profect.tickle.domain.event.service.application;
 
 import com.profect.tickle.domain.event.dto.response.CouponListResponseDto;
 import com.profect.tickle.domain.event.dto.response.CouponResponseDto;
