+++ src/main/java/com/profect/tickle/domain/event/service/application/EventCoreLockService.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.event.service.event;
+package com.profect.tickle.domain.event.service.application;
 
 import com.profect.tickle.domain.event.dto.EventDecision;
 import com.profect.tickle.domain.event.entity.Event;
