+++ src/main/java/com/profect/tickle/domain/event/service/application/impl/EventServiceImpl.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.event.service.event.impl;
+package com.profect.tickle.domain.event.service.application.impl;
 
 import com.profect.tickle.domain.event.dto.request.CouponCreateRequestDto;
 import com.profect.tickle.domain.event.dto.request.TicketEventCreateRequestDto;
@@ -11,11 +11,10 @@ import com.profect.tickle.domain.event.mapper.CouponReceivedMapper;
 import com.profect.tickle.domain.event.mapper.EventMapper;
 import com.profect.tickle.domain.event.repository.CouponRepository;
 import com.profect.tickle.domain.event.repository.EventRepository;
-import com.profect.tickle.domain.event.service.event.EventService;
-import com.profect.tickle.domain.event.service.lock.PessimisticEventApplyExecutor;
+import com.profect.tickle.domain.event.service.application.EventService;
 import com.profect.tickle.domain.event.service.message.publisher.EventPublisher;
-import com.profect.tickle.domain.event.stream.dto.EventMessage;
-import com.profect.tickle.domain.event.stream.producer.EventProducer;
+import com.profect.tickle.domain.event.service.stream.dto.EventMessage;
+import com.profect.tickle.domain.event.service.stream.producer.EventProducer;
 import com.profect.tickle.domain.performance.entity.Performance;
 import com.profect.tickle.domain.performance.repository.PerformanceRepository;
 import com.profect.tickle.domain.point.entity.PointTarget;
@@ -59,7 +58,6 @@ public class EventServiceImpl implements EventService {
     private final CouponReceivedMapper couponReceivedMapper;
     private final PerformanceRepository performanceRepository;
     private final StatusProvider statusProvider;
-    private final EventPublisher eventPublisher;
     private final EventProducer producer;
