+++ src/main/java/com/profect/tickle/domain/event/service/application/impl/PessimisticEventApplyExecutor.java
@@ -1,4 +1,4 @@
-package com.profect.tickle.domain.event.service.lock;
+package com.profect.tickle.domain.event.service.application.impl;
 
 import com.profect.tickle.domain.event.entity.Coupon;
 import com.profect.tickle.domain.event.entity.Event;
@@ -17,7 +17,6 @@ import com.profect.tickle.global.status.Status;
 import com.profect.tickle.global.status.StatusIds;
 import com.profect.tickle.global.status.service.StatusProvider;
 import lombok.RequiredArgsConstructor;
-import org.redisson.api.RedissonClient;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
