--- src/main/java/com/profect/tickle/domain/event/service/message/dto/TicketAppliedDto.java
@@ -1,16 +1,0 @@
-/*
-package com.profect.tickle.domain.event.service.lock;
-
-import com.profect.tickle.domain.point.entity.PointTarget;
-
-public record TicketAppliedDto(
-        Long eventId,
-        Long memberId,
-        Long seatId,
-        short perPrice,
-        int accrued,
-        PointTarget target,
-        boolean isWinner
-) {
-}
-*/
