--- src/main/java/com/profect/tickle/domain/event/service/message/dto/TicketLockMessage.java
@@ -1,20 +1,0 @@
-package com.profect.tickle.domain.event.service.message.dto;
-
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonTypeInfo(
-        use = JsonTypeInfo.Id.CLASS,
-        include = JsonTypeInfo.As.PROPERTY,
-        property = "@class"
-)
-public class TicketLockMessage{
-    Long eventId;
-    Long memberId;
-}
